pyproject.toml
[project]
name = "phone-agent"
version = "0.1.0"
description = "Multi-Provider Phone Agent - Android automation with VLM"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Phone Agent Team" }
]
keywords = ["android", "automation", "vlm", "agent", "adb"]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "textual>=0.89.0",  # TUI 框架
    "rich>=13.0.0",     # 富文本输出
    "pillow>=10.0.0",   # 图像处理
    "aiofiles>=24.0.0", # 异步文件操作
    "adbutils>=2.0.0",  # ADB 操作库
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.scripts]
phone-agent = "phone_agent.tui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
